+++ app/javascript/modules/catalog/mappers/mapCollection.js
@@ -1,8 +1,0 @@
-export default function mapCollection(locale, collection) {
-    return {
-        type: 'collection',
-        id: collection.id,
-        name: collection.name[locale],
-        num_interviews: collection.num_interviews,
-    };
-}