+++ app/javascript/modules/catalog/mappers/mapInterview.js
@@ -1,9 +1,0 @@
-export default function mapInterview(locale, interview) {
-    return {
-        type: 'interview',
-        id: interview.id,
-        shortname: interview.archive_id,
-        name: interview.anonymous_title[locale],
-        project_id: interview.project_id,
-    };
-}