+++ app/javascript/modules/user-agent/index.js
@@ -3,4 +3,5 @@ export { default as ResizeWatcher } from './ResizeWatcher';
 export { default as scrollSmoothlyTo } from './scrollSmoothlyTo';
 export { default as ScrollToTop } from './ScrollToTop';
 export { default as typeOfWithNull } from './typeOfWithNull';
+export { default as useIsMobile } from './useIsMobile';
 export { default as useScrollBelowThreshold } from './useScrollBelowThreshold';