+++ config/initializers/delayed_job.rb
@@ -1,2 +1,2 @@
 Delayed::Worker.max_attempts  = 1
-Delayed::Worker.max_run_time = 15.minutes
+Delayed::Worker.max_run_time = 30.minutes